refactor: unify Client and ClientBrowser file creation
