--- packages/client-generator-ts/src/TSClient/file-generators/ClientBrowserFile.ts
@@ -1,119 +1,0 @@
-import { getClientEngineType } from '@prisma/internals'
-import * as ts from '@prisma/ts-builders'
-
-import { GenerateContext } from '../GenerateContext'
-import { getPrismaClientClassDocComment } from '../PrismaClient'
-import { TSClientOptions } from '../TSClient'
-
-const jsDocHeader = `/*
- * This file should be your main import to use Prisma in the browser. Through it you get access to all the models, enums, and input types.
- * This file however does not contain a functioning PrismaClient class and various helpers are stubbed out as well.
- *
- * 🟢 You can import this file directly.
- */
-`
-
-export function createClientBrowserFile(context: GenerateContext, options: TSClientOptions): string {
-  const clientEngineType = getClientEngineType(options.generator)
-  options.generator.config.engineType = clientEngineType
-
-  const imports = [
-    ts.moduleImport(`${context.runtimeBase}/index-browser`).asNamespace('runtime').typeOnly(),
-    ts.moduleImport(context.importFileName('./enums')).asNamespace('$Enums'),
-    ts.moduleImport(context.importFileName('./internal/class')).asNamespace('$Class').typeOnly(),
-    ts.moduleImport(context.importFileName('./internal/prismaNamespaceBrowser')).asNamespace('Prisma'),
-  ].map((i) => ts.stringify(i))
-
-  const exports = [
-    ts.moduleExportFrom(context.importFileName('./enums')).asNamespace('$Enums'),
-    ts
-      .moduleExport(
-        ts
-          .constDeclaration('PrismaClient')
-          .setValue(
-            ts.namedValue('PrismaClientStub').as(ts.unknownType).as(ts.namedType('$Class.PrismaClientConstructor')),
-          ),
-      )
-      .setDocComment(getPrismaClientClassDocComment(context)),
-    ts.moduleExport(
-      ts
-        .typeDeclaration(
-          'PrismaClient',
-          ts
-            .namedType('$Class.PrismaClient')
-            .addGenericArgument(ts.namedType('LogOpts'))
-            .addGenericArgument(ts.namedType('OmitOpts'))
-            .addGenericArgument(ts.namedType('ExtArgs')),
-        )
-        .addGenericParameter(
-          ts.genericParameter('LogOpts').extends(ts.namedType('Prisma.LogLevel')).default(ts.neverType),
-        )
-        .addGenericParameter(
-          ts
-            .genericParameter('OmitOpts')
-            .extends(ts.namedType('Prisma.PrismaClientOptions').subKey('omit'))
-            .default(ts.namedType('Prisma.PrismaClientOptions').subKey('omit')),
-        )
-        .addGenericParameter(
-          ts
-            .genericParameter('ExtArgs')
-            .extends(ts.namedType('runtime.Types.Extensions.InternalArgs'))
-            .default(ts.namedType('runtime.Types.Extensions.DefaultArgs')),
-        ),
-    ),
-  ].map((e) => ts.stringify(e))
-
-  const modelExports = Object.values(context.dmmf.typeAndModelMap)
-    .filter((model) => context.dmmf.outputTypeMap.model[model.name])
-    .map((model) => {
-      const docLines = model.documentation ?? ''
-      const modelLine = `Model ${model.name}\n`
-      const docs = `${modelLine}${docLines}`
-
-      const modelTypeExport = ts
-        .moduleExport(ts.typeDeclaration(model.name, ts.namedType(`Prisma.${model.name}Model`)))
-        .setDocComment(ts.docComment(docs))
-
-      return ts.stringify(modelTypeExport)
-    })
-
-  const modelEnumsAliases = context.dmmf.datamodel.enums.map((datamodelEnum) => {
-    return [
-      ts.stringify(
-        ts.moduleExport(ts.typeDeclaration(datamodelEnum.name, ts.namedType(`$Enums.${datamodelEnum.name}`))),
-      ),
-      ts.stringify(
-        ts.moduleExport(
-          ts.constDeclaration(datamodelEnum.name).setValue(ts.namedValue(`$Enums.${datamodelEnum.name}`)),
-        ),
-      ),
-    ].join('\n')
-  })
-
-  return `${jsDocHeader}
-${imports.join('\n')}
-
-${stubPrismaClientClass()}
-
-${exports.join('\n')}
-export { Prisma }
-
-${modelExports.join('\n')}
-
-${modelEnumsAliases.length > 0 ? `${modelEnumsAliases.join('\n\n')}` : ''}
-`
-}
-
-function stubPrismaClientClass(): string {
-  return `
-class PrismaClientStub {
-  constructor() {
-    return new Proxy(this, {
-      get(target, prop) {
-        throw new Error('You are referencing a PrismaClient stub defined in <generated-prisma-client>/clientBrowser.ts. Make sure you import the correct <generated-prisma-client>/client.ts file.')
-      }
-    })
-  }
-}
-`
-}
